render: Fix some issues with map iframes.
When loading a page with an embedded map, I see an error
like this in Chrome 92.0.4515.162:

Uncaught DOMException: Blocked a frame with origin "null" from accessing a cross-origin frame.
    at Object._.L.addDomListener (https://maps.googleapis.com/maps/api/js?key=XXX:234:56)
    at HTMLIFrameElement.<anonymous> (https://maps.googleapis.com/maps-api-v3/api/js/46/4/map.js:113:450)
    at ns.i (https://maps.googleapis.com/maps-api-v3/api/js/46/4/map.js:113:501)
    at https://maps.googleapis.com/maps/api/js?key=XXX:155:82

There's another similar error from the Password Alert
extension. I reported this against the Maps API:
https://issuetracker.google.com/issues/199496152

Adding allow-same-origin to the iframe's sandbox attribute
appears to prevent these errors, but it can only be done on
the non-AMP version of the page since amp-iframe doesn't
permit allow-same-origin if the iframe is served from the
same origin as the main page:

https://github.com/ampproject/amphtml/blob/main/docs/spec/amp-iframe-origin-policy.md
https://github.com/ampproject/amphtml/issues/17212

Still, I figure that fixing this for non-AMP pages is better
than nothing.

Also set the iframe's referrerpolicy attribute to
unsafe-url, since using the default
(strict-origin-when-cross-origin) seems to sometimes result
in just the origin getting passed to the iframe, which
breaks marker links on the map. (I see this when testing
locally, at least.)