render: Get rid of margin above initial paragraphs.
Remove the top margin on p:first-of-type within box bodies
in addition to that of *:first-child. The latter matched
paragraphs if they appeared first but not ones that appeared
after left-floated images.

Vertical alignment between text and images appearing further
down in boxes is still a bit off at times, though.