render: Add support for toggling dark mode.
Add an icon to the header that can be used to toggle between
the dark and light themes, and write the setting to
localStorage. When the dark theme is active, the 'dark'
class is set on the <body> element.

Graph iframes still need to be updated to respond to theme
changes.