render: Update syntax highlighting brightness.
Brighten colors in the dark theme and darken them in the
light theme so that Lighthouse accessibility audits
pass (at least when using the lovelace and paraiso-dark
styles).

Also add some mappings to handle token types that are
missing in the dark style, and update a CSS selector to
prevent the light style's rules from being used as a
fallback (even if the colors are overridden by the dark
style, underlining, italics, etc. won't be).